collaborative-comfyui/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
python-socketio==5.10.0
websockets==11.0.3
aiohttp==3.8.6
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
alembic==1.12.1
email-validator==2.1.0
